+++ cmd/template/nodepool/provider/templates/aws/nodepool.go
@@ -4,11 +4,12 @@ import (
 	"github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha3"
 	"github.com/giantswarm/k8smetadata/pkg/annotation"
 	"github.com/giantswarm/k8smetadata/pkg/label"
+	"github.com/giantswarm/microerror"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	apiv1alpha3 "sigs.k8s.io/cluster-api/api/v1alpha3"
 
-	"github.com/giantswarm/kubectl-gs/pkg/id"
+	"github.com/giantswarm/kubectl-gs/internal/key"
 )
 
 const (
@@ -20,9 +21,10 @@ const (
 type NodePoolCRsConfig struct {
 	AvailabilityZones                   []string
 	AWSInstanceType                     string
-	ClusterID                           string
-	MachineDeploymentID                 string
+	ClusterName                         string
+	MachineDeploymentName               string
 	Description                         string
+	EnableLongNames                     bool
 	NodesMax                            int
 	NodesMin                            int
 	OnDemandBaseCapacity                int
@@ -44,11 +46,22 @@ func NewNodePoolCRs(config NodePoolCRsConfig) (NodePoolCRs, error) {
 	// Default some essentials in case certain information are not given. E.g.
 	// the workload cluster ID may be provided by the user.
 	{
-		if config.ClusterID == "" {
-			config.ClusterID = id.Generate()
+		if config.ClusterName == "" {
+			generatedName, err := key.GenerateName(config.EnableLongNames)
+			if err != nil {
+				return NodePoolCRs{}, microerror.Mask(err)
+			}
+
+			config.ClusterName = generatedName
 		}
-		if config.MachineDeploymentID == "" {
-			config.MachineDeploymentID = id.Generate()
+
+		if config.MachineDeploymentName == "" {
+			generatedName, err := key.GenerateName(config.EnableLongNames)
+			if err != nil {
+				return NodePoolCRs{}, microerror.Mask(err)
+			}
+
+			config.MachineDeploymentName = generatedName
 		}
 	}
 
@@ -70,18 +83,18 @@ func newAWSMachineDeploymentCR(c NodePoolCRsConfig) *v1alpha3.AWSMachineDeployme
 			APIVersion: v1alpha3.SchemeGroupVersion.String(),
 		},
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      c.MachineDeploymentID,
+			Name:      c.MachineDeploymentName,
 			Namespace: metav1.NamespaceDefault,
 			Annotations: map[string]string{
 				annotation.Docs: "https://docs.giantswarm.io/ui-api/management-api/crd/awsmachinedeployments.infrastructure.giantswarm.io/",
 			},
 			Labels: map[string]string{
 				label.AWSOperatorVersion:     c.ReleaseComponents["aws-operator"],
-				label.Cluster:                c.ClusterID,
-				label.MachineDeployment:      c.MachineDeploymentID,
+				label.Cluster:                c.ClusterName,
+				label.MachineDeployment:      c.MachineDeploymentName,
 				label.Organization:           c.Owner,
 				label.ReleaseVersion:         c.ReleaseVersion,
-				apiv1alpha3.ClusterLabelName: c.ClusterID,
+				apiv1alpha3.ClusterLabelName: c.ClusterName,
 			},
 		},
 		Spec: v1alpha3.AWSMachineDeploymentSpec{
@@ -118,25 +131,25 @@ func newMachineDeploymentCR(obj *v1alpha3.AWSMachineDeployment, c NodePoolCRsCon
 			APIVersion: "cluster.x-k8s.io/v1alpha3",
 		},
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      c.MachineDeploymentID,
+			Name:      c.MachineDeploymentName,
 			Namespace: metav1.NamespaceDefault,
 			Annotations: map[string]string{
 				annotation.Docs: "https://docs.giantswarm.io/ui-api/management-api/crd/machinedeployments.cluster.x-k8s.io/",
 			},
 			Labels: map[string]string{
-				label.Cluster:                c.ClusterID,
+				label.Cluster:                c.ClusterName,
 				label.ClusterOperatorVersion: c.ReleaseComponents["cluster-operator"],
-				label.MachineDeployment:      c.MachineDeploymentID,
+				label.MachineDeployment:      c.MachineDeploymentName,
 				label.Organization:           c.Owner,
 				label.ReleaseVersion:         c.ReleaseVersion,
-				apiv1alpha3.ClusterLabelName: c.ClusterID,
+				apiv1alpha3.ClusterLabelName: c.ClusterName,
 			},
 		},
 		Spec: apiv1alpha3.MachineDeploymentSpec{
-			ClusterName: c.ClusterID,
+			ClusterName: c.ClusterName,
 			Template: apiv1alpha3.MachineTemplateSpec{
 				Spec: apiv1alpha3.MachineSpec{
-					ClusterName: c.ClusterID,
+					ClusterName: c.ClusterName,
 					InfrastructureRef: corev1.ObjectReference{
 						APIVersion: obj.TypeMeta.APIVersion,
 						Kind:       obj.TypeMeta.Kind,